refactor(experiments): enforce the page-change rule in the API, not the REST layer
@freddyDOTCMS asked for this on #37206 and the automated review reached the
same place independently.

The rule and the control-url regeneration were split across two layers: the
resource validated, save() regenerated. That left save() half-guarded — a
caller reaching experimentsAPI.save() directly with a changed pageId on a
non-draft experiment, or one carrying real variants, would move page_id while
leaving those variants' urls pointing at the old page.

The rule now sits next to the regeneration it protects, so both run for every
caller. Only a changed pageId reaches the check, so saves that keep their page
(addVariant, deleteVariant, promoteVariant, start, end) pay nothing.

The spec is unaffected: FR-002 never said which layer enforces the rule. If
anything this is closer to it, since the previous version let a non-REST caller
violate FR-002 outright.

All 8 integration tests still pass — they assert behaviour, not placement.

Refs #37176
